+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- TODO: Describe initial release.
+
+## 1.0.1
+
+- implement data transfer.
+
+## 1.0.2
+
+- implement permission handler